flask_server/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
openai==1.57.2
python-dotenv==1.0.1
